requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
PyQt6>=6.4.0
pyautogui>=0.9.53
pillow>=9.3.0
rumps==0.4.0
requests>=2.31.0
pyautogui==0.9.54
